[FIX] website: ignore mutations on hide/show a modal popup in builder
Some of the mutations happening when toggling the visibility of a popup
were tracked in the history. Thus, when toggling the visibility of a
popup and doing undo/redo, the actual visibility of a popup could be
inconsistent with the state of the toggle button.

To avoid adding mutations in the history when toggling visibility, using
`ignoreDOMMutations` around `hide` and `show` is not enough because
Bootstrap delays some operations

task-4367641